:white_check_mark: added a stress test for 250VU. Check commit description for further analysis.
Test maded in a M2 16GB RAM MacBook. For 7881 requisitions, none of them failed.Average latency of 741ms. P95 of 4.83s with a max of 5.85s. CPU resources with a average latency of 2.79ms. Conclusion: the test passed successfully, but its noticeble a linear increase in response time, resulting in a small delay for 5% of the virtual users.